making layers check cookies
