+++ .gitignore
@@ -74,11 +74,11 @@ web_modules/
 .yarn-integrity
 
 # dotenv environment variable files
-#.env
-#.env.development.local
-#.env.test.local
-#.env.production.local
-#.env.local
+.env
+.env.development.local
+.env.test.local
+.env.production.local
+.env.local
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
